+++ src/components/sections/elements/__tests__/Citation.spec.ts
@@ -104,8 +104,11 @@ describe('Citation', () => {
     // getCitation() is async, so need to wait for it to resolve otherwise element will be empty
     await flushPromises()
 
-    wrapper.find('button').trigger('click')
+    wrapper.find('button#citation-use-generated').trigger('click')
     await wrapper.vm.$nextTick()
+
+    // dump contents
+    console.log(wrapper.find('textarea').element.value)
 
     expect(wrapper.find('textarea').element.value).toContain(`_${record.title}_`)
   })